add optimization for MemoryStorageStep
