+++ src/Storages/StorageMemory.cpp
@@ -9,15 +9,13 @@
 #include <Storages/StorageMemory.h>
 #include <Storages/MemorySettings.h>
 #include <DataTypes/ObjectUtils.h>
-#include <Columns/ColumnObject.h>
 
 #include <IO/WriteHelpers.h>
-#include <Processors/ISource.h>
 #include <QueryPipeline/Pipe.h>
 #include <QueryPipeline/QueryPipelineBuilder.h>
 #include <Processors/Sinks/SinkToStorage.h>
 #include <Processors/Executors/PullingPipelineExecutor.h>
-#include <Processors/QueryPlan/ISourceStep.h>
+#include <Processors/QueryPlan/ReadFromMemoryStorageStep.h>
 #include <Parsers/ASTCreateQuery.h>
 
 #include <Common/FileChecker.h>
@@ -43,85 +41,6 @@ namespace ErrorCodes
     extern const int NUMBER_OF_ARGUMENTS_DOESNT_MATCH;
     extern const int CANNOT_RESTORE_TABLE;
 }
-
-
-class MemorySource : public ISource
-{
-    using InitializerFunc = std::function<void(std::shared_ptr<const Blocks> &)>;
-public:
-
-    MemorySource(
-        Names column_names_,
-        const StorageSnapshotPtr & storage_snapshot,
-        std::shared_ptr<const Blocks> data_,
-        std::shared_ptr<std::atomic<size_t>> parallel_execution_index_,
-        InitializerFunc initializer_func_ = {})
-        : ISource(storage_snapshot->getSampleBlockForColumns(column_names_))
-        , column_names_and_types(storage_snapshot->getColumnsByNames(
-            GetColumnsOptions(GetColumnsOptions::All).withSubcolumns().withExtendedObjects(), column_names_))
-        , data(data_)
-        , parallel_execution_index(parallel_execution_index_)
-        , initializer_func(std::move(initializer_func_))
-    {
-    }
-
-    String getName() const override { return "Memory"; }
-
-protected:
-    Chunk generate() override
-    {
-        if (initializer_func)
-        {
-            initializer_func(data);
-            initializer_func = {};
-        }
-
-        size_t current_index = getAndIncrementExecutionIndex();
-
-        if (!data || current_index >= data->size())
-        {
-            return {};
-        }
-
-        const Block & src = (*data)[current_index];
-
-        Columns columns;
-        size_t num_columns = column_names_and_types.size();
-        columns.reserve(num_columns);
-
-        auto name_and_type = column_names_and_types.begin();
-        for (size_t i = 0; i < num_columns; ++i)
-        {
-            columns.emplace_back(tryGetColumnFromBlock(src, *name_and_type));
-            ++name_and_type;
-        }
-
-        fillMissingColumns(columns, src.rows(), column_names_and_types, column_names_and_types, {}, nullptr);
-        assert(std::all_of(columns.begin(), columns.end(), [](const auto & column) { return column != nullptr; }));
-
-        return Chunk(std::move(columns), src.rows());
-    }
-
-private:
-    size_t getAndIncrementExecutionIndex()
-    {
-        if (parallel_execution_index)
-        {
-            return (*parallel_execution_index)++;
-        }
-        else
-        {
-            return execution_index++;
-        }
-    }
-
-    const NamesAndTypesList column_names_and_types;
-    size_t execution_index = 0;
-    std::shared_ptr<const Blocks> data;
-    std::shared_ptr<std::atomic<size_t>> parallel_execution_index;
-    InitializerFunc initializer_func;
-};
-
 
 class MemorySink : public SinkToStorage
 {
@@ -190,83 +109,6 @@ private:
 
     StorageMemory & storage;
     StorageSnapshotPtr storage_snapshot;
-};
-
-
-class ReadFromMemoryStorageStep final : public ISourceStep
-{
-public:
-    explicit ReadFromMemoryStorageStep(Pipe pipe_) :
-        ISourceStep(DataStream{.header = pipe_.getHeader()}),
-        pipe(std::move(pipe_))
-    {
-    }
-
-    ReadFromMemoryStorageStep() = delete;
-    ReadFromMemoryStorageStep(const ReadFromMemoryStorageStep &) = delete;
-    ReadFromMemoryStorageStep & operator=(const ReadFromMemoryStorageStep &) = delete;
-
-    ReadFromMemoryStorageStep(ReadFromMemoryStorageStep &&) = default;
-    ReadFromMemoryStorageStep & operator=(ReadFromMemoryStorageStep &&) = default;
-
-    String getName() const override { return name; }
-
-    void initializePipeline(QueryPipelineBuilder & pipeline, const BuildQueryPipelineSettings &) override
-    {
-        // use move - make sure that the call will only be made once.
-        pipeline.init(std::move(pipe));
-    }
-
-    static Pipe makePipe(const Names & columns_to_read_,
-                         const StorageSnapshotPtr & storage_snapshot_,
-                         size_t num_streams_,
-                         const bool delay_read_for_global_subqueries_)
-    {
-        storage_snapshot_->check(columns_to_read_);
-
-        const auto & snapshot_data = assert_cast<const StorageMemory::SnapshotData &>(*storage_snapshot_->data);
-        auto current_data = snapshot_data.blocks;
-
-        if (delay_read_for_global_subqueries_)
-        {
-            /// Note: for global subquery we use single source.
-            /// Mainly, the reason is that at this point table is empty,
-            /// and we don't know the number of blocks are going to be inserted into it.
-            ///
-            /// It may seem to be not optimal, but actually data from such table is used to fill
-            /// set for IN or hash table for JOIN, which can't be done concurrently.
-            /// Since no other manipulation with data is done, multiple sources shouldn't give any profit.
-
-            return Pipe(std::make_shared<MemorySource>(
-                columns_to_read_,
-                storage_snapshot_,
-                nullptr /* data */,
-                nullptr /* parallel execution index */,
-                [current_data](std::shared_ptr<const Blocks> & data_to_initialize)
-                {
-                    data_to_initialize = current_data;
-                }));
-        }
-
-        size_t size = current_data->size();
-
-        if (num_streams_ > size)
-            num_streams_ = size;
-
-        Pipes pipes;
-
-        auto parallel_execution_index = std::make_shared<std::atomic<size_t>>(0);
-
-        for (size_t stream = 0; stream < num_streams_; ++stream)
-        {
-            pipes.emplace_back(std::make_shared<MemorySource>(columns_to_read_, storage_snapshot_, current_data, parallel_execution_index));
-        }
-        return Pipe::unitePipes(std::move(pipes));
-    }
-
-private:
-    static constexpr auto name = "ReadFromMemoryStorage";
-    Pipe pipe;
 };
