+++ src/Storages/StorageMemory.h
@@ -5,6 +5,7 @@
 #include <mutex>
 
 #include <Core/NamesAndTypes.h>
+#include <Interpreters/DatabaseCatalog.h>
 #include <Storages/IStorage.h>
 
 #include <Common/MultiVersion.h>
